Remove Note.advanced(by:) to avoid confusion
This method suggests that Note could conform to Strideable, which it
cannot (without violating the protocol's laws) due to enharmonic
equivalents, which would be considered different according to Equatable
but equivalent according to Comparable.

To advance a note by a number of semitones without regard to the chosen
enharmonic equivalent, use the + operator with a ChromaticInterval.